pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Console gendiff utility"
authors = ["skip"]
readme = "README.md"
license = "MIT"
packages = [{include = "gendiff"}]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console"
]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
